update k8s manifest

+++ script/pushToACR.sh
@@ -6,7 +6,9 @@ az acr login --name $acr
 
 # Remove image if exist
 existing_image=$(docker images | grep sampleapp | awk '{print $1}')
-docker rmi $existing_image
+if [[ ! -z "$existing_image" ]]
+then
+    docker rmi $existing_image
 
 docker build -t $acr/dotnet/sampleapp:latest ../aspnet-core-dotnet-core/
 docker push $acr/dotnet/sampleapp:latest
